requirements.txt
PyYAML>=5.3.1
pandas>=1.1.0
numpy>=1.19.0
matplotlib>=3.2.0
seaborn>=0.10.0
scipy>=1.5.0
scikit-learn>=0.23.0
umap-learn>=0.4.6
hdbscan>=0.8.27
networkx>=2.5
